+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Block Injector Changelog
 
+## Unreleased
+
+### Added
+
+- `atIndex` method for 0-based injection
+
+### Changed
+
+- Fixed interval behavior
+- Fixed issues with negative index injection
+
+### Removed
+
+- `intervalCallback` as it didn't really make sense…
+
 ## 0.1.0 - 2020-10-12
 
 ### Added